requirements.txt
numpy>=1.21.0
Pillow>=9.0.0
PyWavelets>=1.1.1
matplotlib>=3.4.0
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
scipy==1.10.1 